move common kafka bits to commmon module, add log_entries
