+++ rust/common/kafka/src/kafka_messages/app_metrics2.rs
