+++ rust/common/kafka/src/kafka_messages/mod.rs
@@ -1,5 +1,6 @@
 pub mod app_metrics;
 pub mod app_metrics2;
+pub mod log_entries;
 pub mod plugin_logs;
 
 use chrono::{DateTime, NaiveDateTime, Utc};
