+++ rust/common/kafka/src/kafka_messages/plugin_logs.rs
