+++ rust/common/kafka/src/kafka_producer.rs
