+++ rust/common/kafka/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod config;
+pub mod kafka_messages;
+pub mod kafka_producer;
+pub mod test;
