+++ rust/hook-common/src/lib.rs
@@ -1,7 +1,3 @@
-pub mod config;
-pub mod kafka_messages;
-pub mod kafka_producer;
 pub mod pgqueue;
 pub mod retry;
-pub mod test;
 pub mod webhook;
