+++ rust/hook-janitor/Cargo.toml
@@ -25,3 +25,4 @@ tokio = { workspace = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 common-metrics = { path = "../common/metrics" }
+common-kafka = { path = "../common/kafka" }
