+++ rust/hook-janitor/src/main.rs
@@ -9,8 +9,8 @@ use std::{str::FromStr, time::Duration};
 use tokio::sync::Semaphore;
 use webhooks::WebhookCleaner;
 
+use common_kafka::kafka_producer::create_kafka_producer;
 use common_metrics::setup_metrics_routes;
-use hook_common::kafka_producer::create_kafka_producer;
 
 mod cleanup;
 mod config;
