+++ rust/hook-worker/Cargo.toml
@@ -26,7 +26,8 @@ tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 url = { version = "2.2" }
 common-metrics = { path = "../common/metrics" }
-common-dns = { path =  "../common/dns" }
+common-dns = { path = "../common/dns" }
+common-kafka = { path = "../common/kafka" }
 
 [dev-dependencies]
 httpmock = { workspace = true }
