+++ rust/hook-worker/src/worker.rs
@@ -15,7 +15,7 @@ use tokio::sync;
 use tokio::time::{sleep, Duration};
 use tracing::error;
 
-use hook_common::kafka_producer::KafkaContext;
+use common_kafka::kafka_producer::KafkaContext;
 use hook_common::pgqueue::PgTransactionBatch;
 use hook_common::{
     pgqueue::{Job, PgQueue, PgQueueJob, PgTransactionJob, RetryError, RetryInvalidError},
@@ -793,9 +793,9 @@ mod tests {
     // Note we are ignoring some warnings in this module.
     // This is due to a long-standing cargo bug that reports imports and helper functions as unused.
     // See: https://github.com/rust-lang/rust/issues/46379.
+    use common_kafka::test::create_mock_kafka;
     use health::HealthRegistry;
     use hook_common::pgqueue::{DatabaseError, NewJob};
-    use hook_common::test::create_mock_kafka;
     use hook_common::webhook::WebhookJobMetadata;
     use sqlx::PgPool;
